[os_crypto] make some casts explicit

+++ src/os_crypto/md5/md5_op.c
@@ -29,7 +29,7 @@ int OS_MD5_File(const char * fname, os_md5 output)
     MD5_CTX ctx;
     unsigned char buf[1024 +1];
     unsigned char digest[16];
-    int n;
+    size_t n;
 
     memset(output,0, 33);
     buf[1024] = '\0';
@@ -44,7 +44,7 @@ int OS_MD5_File(const char * fname, os_md5 output)
     while((n = fread(buf, 1, sizeof(buf) -1, fp)) > 0)
     {
         buf[n] = '\0';
-        MD5Update(&ctx,buf,n);
+        MD5Update(&ctx,buf,(unsigned)n);
     }
 
     MD5Final(digest, &ctx);
@@ -71,7 +71,7 @@ int OS_MD5_Str(const char * str, os_md5 output)
 
     MD5Init(&ctx);
 
-    MD5Update(&ctx,(unsigned char *)str,strlen(str));
+    MD5Update(&ctx,(unsigned char *)str,(unsigned)strlen(str));
 
     MD5Final(digest, &ctx);
